Move data.py and update API doc.
